+++ plone/recipe/codeanalysis/python_utf8_header.py
@@ -1,32 +1,35 @@
 # -*- coding: utf-8 -*-
-from plone.recipe.codeanalysis.utils import find_files
-from plone.recipe.codeanalysis.utils import log
+from plone.recipe.codeanalysis.analyser import console_factory
+from plone.recipe.codeanalysis.clean_lines import CleanLines
+import re
 
 
-def code_analysis_utf8_header(options):
-    log('title', 'Check utf-8 headers')
+class UTF8Headers(CleanLines):
 
-    files = find_files(options, '.*\.py')
-    if not files:
-        log('ok')
-        return True
+    name = 'utf8-header'
+    title = 'Check utf-8 headers'
+    message = '{0:s}: missing utf-8 header'
+    checks = [
+        {
+            'extensions': ('py', ),
+            'succeed': r'^#\s-\*-\scoding:\sutf-8\s-\*-$',
+        },
+    ]
 
-    errors = []
-    file_paths = files.strip().split('\n')
-    for file_path in file_paths:
-        with open(file_path, 'r') as file_handler:
+    @staticmethod
+    def validate_line(pattern, line):
+        return re.compile(pattern).match(line)
 
-            lines = file_handler.readlines()
-            if len(lines) == 0:
-                continue
-            elif lines[0].find('coding: utf-8') == -1:
-                errors.append('{0}: missing utf-8 header'.format(file_path))
+    def check(self, file_path, succeed, **kwargs):
+        with open(file_path, 'r') as file_handle:
 
-    if len(errors) > 0:
-        log('failure')
-        for err in errors:
-            print(err)
-        return False
-    else:
-        log('ok')
-        return True
+            # Just check the first line of each file
+            line = file_handle.readline()
+            if UTF8Headers.validate_line(succeed, line):
+                return []
+
+        return [self.message.format(file_path), ]
+
+
+def console_script(options):
+    console_factory(UTF8Headers, options)