+++ plone/recipe/codeanalysis/tests/test_jscs.py
@@ -3,9 +3,9 @@ from __future__ import unicode_literals
 from io import open
 from os.path import isfile as path_isfile
 from os.path import join as path_join
-from plone.recipe.codeanalysis.jscs import code_analysis_jscs
-from plone.recipe.codeanalysis.jscs import jscs_errors
+from plone.recipe.codeanalysis.jscs import JSCS
 from shutil import rmtree
+from tempfile import TemporaryFile
 from tempfile import mkdtemp
 import unittest
 
@@ -77,7 +77,7 @@ class TestJavascriptCodeStyleChecker(unittest.TestCase):
         with open(full_path, 'w') as incorrect_code:
             incorrect_code.write(INCORRECT_FILE)
         self.options['directory'] = self.test_dir
-        self.assertFalse(code_analysis_jscs(self.options))
+        self.assertFalse(JSCS(self.options).run())
 
     def test_analysis_should_return_true_when_invalid_file_is_excluded(self):
         full_path = path_join(self.test_dir, 'incorrect.js')
@@ -85,21 +85,24 @@ class TestJavascriptCodeStyleChecker(unittest.TestCase):
             incorrect_code.write(INCORRECT_FILE)
         self.options['directory'] = self.test_dir
         self.options['jscs-exclude'] = full_path
-        self.assertTrue(code_analysis_jscs(self.options))
+        self.assertTrue(JSCS(self.options).run())
 
     def test_analysis_should_return_false_when_oserror(self):
+        full_path = path_join(self.test_dir, 'incorrect.js')
+        with open(full_path, 'w') as incorrect_code:
+            incorrect_code.write(INCORRECT_FILE)
+        self.options['directory'] = self.test_dir
         # The options are fake, so the function should raise an OSError
         # and return false.
         self.options['jscs-bin'] = 'FAKE_BIN'
-        self.options['directory'] = 'FAKE_DIR'
-        self.assertFalse(code_analysis_jscs(self.options))
+        self.assertFalse(JSCS(self.options).run())
 
     def test_analysis_should_return_true(self):
         full_path = path_join(self.test_dir, 'correct.js')
         with open(full_path, 'w') as correct_code:
             correct_code.write(CORRECT_FILE)
         self.options['directory'] = self.test_dir
-        self.assertTrue(code_analysis_jscs(self.options))
+        self.assertTrue(JSCS(self.options).run())
 
     def test_analysis_file_should_exist_when_jenkins_is_true(self):
         location_tmp_dir = mkdtemp()
@@ -109,19 +112,43 @@ class TestJavascriptCodeStyleChecker(unittest.TestCase):
         self.options['directory'] = self.test_dir
         self.options['location'] = location_tmp_dir
         self.options['jenkins'] = 'True'  # need to activate jenkins.
-        code_analysis_jscs(self.options)
+        JSCS(self.options).run()
         file_exist = path_isfile(path_join(location_tmp_dir, 'jscs.xml'))
         rmtree(location_tmp_dir)
         self.assertTrue(file_exist)
 
-    def test_jscs_errors_should_return_false_empty_xml_output(self):
-        self.assertFalse(jscs_errors(XML_EMPTY_OUTPUT, True))
+    def test_jscs_parse_output_should_return_true_empty_xml_output(self):
+        jscs_file = TemporaryFile('w+')
+        jscs_file.write(XML_EMPTY_OUTPUT)
+        jscs_file.seek(0)
+        self.options['jenkins'] = 'True'
+        linter = JSCS(self.options)
+        self.assertTrue(linter.use_jenkins)
+        self.assertTrue(linter.parse_output(jscs_file, 1))
 
-    def test_jscs_errors_should_return_true_with_xml_output(self):
-        self.assertTrue(jscs_errors(XML_OUTPUT, True))
+    def test_jscs_parse_output_should_return_false_with_xml_output(self):
+        jscs_file = TemporaryFile('w+')
+        jscs_file.write(XML_OUTPUT)
+        jscs_file.seek(0)
+        self.options['jenkins'] = 'True'
+        linter = JSCS(self.options)
+        self.assertTrue(linter.use_jenkins)
+        self.assertFalse(linter.parse_output(jscs_file, 1))
 
-    def test_jscs_errors_should_return_true_with_normal_output(self):
-        self.assertTrue(jscs_errors(DEFAULT_OUTPUT, False))
+    def test_jscs_parse_output_should_return_false_with_normal_output(self):
+        jscs_file = TemporaryFile('w+')
+        jscs_file.write(DEFAULT_OUTPUT)
+        jscs_file.seek(0)
+        self.options['jenkins'] = 'False'
+        linter = JSCS(self.options)
+        self.assertFalse(linter.use_jenkins)
+        self.assertFalse(linter.parse_output(jscs_file, 1))
 
-    def test_jscs_errors_should_return_false_empty_normal_output(self):
-        self.assertFalse(jscs_errors('', False))
+    def test_jscs_parse_output_should_return_true_empty_normal_output(self):
+        jscs_file = TemporaryFile('w+')
+        jscs_file.write('')
+        jscs_file.seek(0)
+        self.options['jenkins'] = 'False'
+        linter = JSCS(self.options)
+        self.assertFalse(linter.use_jenkins)
+        self.assertTrue(linter.parse_output(jscs_file, 1))